core: fix stacksize check in thread_create()
- use `size_t` for `stacksize` to not have to worry about negative
  values being passed for it
- also take thread local storage requirements into account when
  verifying stack size
  - the minimum check takes potential loss due to aligning and the
    RIOT data structures into account.
  - it would still be possible to write a thread in assembly to use no
    stack at all and give it just the amount of stack needed to RIOT's
    internal data structure (+ 8 bytes of for possible alignment
    loss on 32 bit systems)

+++ core/include/thread.h
@@ -275,7 +275,7 @@ struct _thread {
  */
 ACCESS(write_only, 1, 2)
 kernel_pid_t thread_create(char *stack,
-                           int stacksize,
+                           size_t stacksize,
                            uint8_t priority,
                            int flags,
                            thread_task_func_t task_func,
